% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_waic_ideal.R
\name{cal_waic_ideal}
\alias{cal_waic_ideal}
\title{Calculate WAIC of the IDEAL model}
\usage{
cal_waic_ideal(chain_results, vote_info, block = F, block_vote = F)
}
\arguments{
\item{chain_results}{Posterior samples obtained from function 'ideal' in 'pscl' package.}

\item{vote_info}{Rollcall object used as input of the 'ideal' function.}
}
\value{
The WAIC value of IDEAL model
}
\description{
This function is used to get the WAIC value of the IDEAL model
}
\examples{
waic_ideal = cal_waic_ideal(post_samples_ideal, h116)
}
