% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_waic_pum_static.R
\name{cal_waic_pum_static}
\alias{cal_waic_pum_static}
\title{Calculate WAIC of the static probit unfolding model}
\usage{
cal_waic_pum_static(chain_run)
}
\arguments{
\item{chain_run}{The object output from 'sample_pum_static' in 'pumBayes'.}
}
\value{
The WAIC value of the static probit unfoldinng model
}
\description{
This function is used to get the WAIC value of the static probit unfolding model
}
\examples{
waic_pum_static = cal_waic_pum_static(chain_run = post_samples_pum)
}
