% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_waic_irt_dynamic.R
\name{calc_waic_irt_dynamic}
\alias{calc_waic_irt_dynamic}
\title{Calculate WAIC of the dynamic ideal point model}
\usage{
calc_waic_irt_dynamic(case_vote_m, case_year_v, chain_results)
}
\arguments{
\item{vote_m}{A vote matrix in which rows represent members and columns represent issues.
The entries can only be 0 (indicating ”No”), 1 (indicating ”Yes”), or NA (indicating missing data).}

\item{years_v}{A vector representing the time period for each vote in the model}

\item{chain_run}{The posterior samples obtained from `MCMCdynamicIRT1d` in `MCMCpack`}
}
\description{
This function is used to get the WAIC value of the dynamic ideal point model prorosed by Martin and Quinn
}
\examples{
waic_irt_dy = calc_waic_irt_dynamic(mqVotes, mqTime, post_samples_irt_dy)
}
