% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pum_dynamic.R, R/sample_pum_static.R
\name{init_y_star_m}
\alias{init_y_star_m}
\title{initialize three auxiliary parameters y}
\usage{
init_y_star_m(vote_m)

init_y_star_m(vote_m)
}
\description{
initialize three auxiliary parameters y

initialize three auxiliary parameters y
}
