% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_rank.R
\name{median_rank}
\alias{median_rank}
\title{Get median rank plots}
\usage{
median_rank(beta, col_blue = "D")
}
\arguments{
\item{beta}{A matrix of posterior samples of beta obtained from MCMC}

\item{col_blue}{A character string indicating whether the "D" party should be colored blue or "R".}
}
\value{
A list containing a ggplot object of the rank plot and a data frame of sorted rank data.
}
\description{
Generates the median rank plot for all members based on posterior samples
}
\examples{
rank_results = median_rank(beta = post_samples$beta, col_blue = "D")
}
