% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_waic.R
\name{calc_waic}
\alias{calc_waic}
\title{Calculate a block version of Watanabe-Akaike Information Criterion (WAIC)}
\usage{
calc_waic(vote_info, years_v = NULL, prob_array)
}
\arguments{
\item{vote_info}{A logical vote matrix (or a rollcall object) in which rows represent members and columns represent issues.
The entries should be FALSE ("No"), TRUE ("Yes"), or NA (missing data).}

\item{years_v}{A vector representing the time period for each vote in the model. This is defultly set as `NULL` for a static model.}

\item{prob_array}{An array of probabilities with three dimensions.}
}
\value{
The block WAIC value for a static PUM or a vector of WAIC by time for a dynamic PUM.
}
\description{
This function is used to get the WAIC value when blocking members
}
\examples{
waic_pum_dy = calc_waic(prob_pum_dy, years_v = NULL, h116.c)
}
