% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_char.R
\name{item_char}
\alias{item_char}
\title{Generate Data for Item Characteristic Curves}
\usage{
item_char(vote_num, x = NULL, post_samples)
}
\arguments{
\item{vote_num}{The vote number of the issue to be reviewed. This refers to numbers
in the column names of the input vote matrix, not the clerk session vote number.}

\item{x}{A vector showing the range of beta in the x axis.}

\item{post_samples}{A list of posterior samples of parameters obtained from `sample_pum_static` in `pumBayes`.}
}
\value{
A data frame containing `beta_samples`, mean probabilities (`means`),
and confidence intervals (`ci_lower` and `ci_upper`) for the input issue,
which can be used to plot the item characteristic curve.
}
\description{
This function calculates the data needed to plot the item characteristic
curve for a specific issue based on posterior samples.
}
\examples{
item_data <- item_char(vote_num = 5, post_samples = post_samples_pum)
}
