% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_rank.R
\name{post_rank}
\alias{post_rank}
\title{Generate Quantile Ranks for Legislators}
\usage{
post_rank(beta, quantiles = c(0.5))
}
\arguments{
\item{beta}{A matrix of posterior samples of beta obtained from MCMC, with columns representing legislators.}

\item{quantiles}{A numeric vector specifying the quantiles to be calculated for the ranks (default is `c(0.5)` for median rank).}
}
\value{
A data frame containing the legislators' names, party affiliations, states, and their ranks at each specified quantile.
If the median is included, it will be named `median` in the output. The output data frame is sorted in ascending order based on the values in the median column.
}
\description{
This function calculates quantile ranks for each legislator based on posterior samples of beta parameters from MCMC.
The function can handle any specified quantiles, such as median (0.5), and is flexible to support other quantiles provided as input.
}
\examples{
rank_results = post_rank(beta = post_samples_pum$beta, quantiles = c(0.5))
}
