% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_irt.R
\name{predict_irt}
\alias{predict_irt}
\title{Calculate Probabilities for Dynamic Item Response Theory Model}
\usage{
predict_irt(vote_info, years_v, post_samples)
}
\arguments{
\item{vote_info}{A logical vote matrix where rows represent members and columns represent issues.
The entries should be FALSE ("No"), TRUE ("Yes"), or NA (missing data).}

\item{years_v}{A vector representing the time period for each vote in the model.}

\item{post_samples}{MCMC results obtained from `wnominate' function in `wnominate' package.}
}
\description{
This function computes the probability matrix for a dynamic item response theory (IRT) model.
Specifically, it calculates the probabilities of voting "Yea" for each legislator (member),
issue, and time period based on the posterior samples of model parameters.
}
\examples{
prob_irt_dy = predict_irt(post_samples_irt_dy, mqVotes, mqTime)
}
